+++ lib/services/carreras_service.dart
@@ -35,8 +35,19 @@ class CarreraService {
     );
 
     List<Carrera> carreras = Carrera.fromJsonList(response.data);
-    Carrera activa = carreras
-        .firstWhere((carrera) => carrera.estado!.toLowerCase() == "regular");
+
+    final estados = ["Regular", "Causal de Eliminacion"]
+        .reversed
+        .map((e) => e.toLowerCase())
+        .toList();
+
+    carreras.sort(
+      (a, b) => estados.indexOf(b.estado!.toLowerCase()).compareTo(
+            estados.indexOf(a.estado!.toLowerCase()),
+          ),
+    );
+
+    Carrera activa = carreras.first;
 
     return activa;
   }