+++ lib/services/notificaciones_service.dart
@@ -37,15 +37,11 @@ class NotificationsService {
         debug: true,
       );
 
-      // Get initial notification action is optional
-      _initialAction = await AwesomeNotifications()
-          .getInitialNotificationAction(removeFromActionEvents: false);
-
       NotificationsService.notification
           .setListeners(onActionReceivedMethod: _onActionReceivedMethod);
 
       String? fcmToken = await FirebaseMessaging.instance.getToken();
-      print("Firebase token: $fcmToken");
+      print("Firebase Cloud Messaging token: $fcmToken");
     } catch (e) {
       print('Error ${e.toString()}');
     }